requirements.txt
Flask==3.0.2
pycryptodome==3.20.0
PyJWT==2.8.0
python-dotenv==1.0.1
cryptography
